% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-alabama.R
\docType{data}
\name{alabama}
\alias{alabama}
\title{Sample of 2017 ACS PUMS data for Alabama}
\format{
A data.frame object with 5000 rows and 286 variables. The object
also inherits the \code{folk_df} class.
}
\usage{
alabama
}
\description{
The data is a random sample of 5000 rows of the 2017 ACS PUMS 1-Year person
survey data.
}
\details{
The original data is available at:
https://www2.census.gov/programs-surveys/acs/data/pums/2017/1-Year/csv_pal.zip
}
\examples{
alabama
}
\keyword{datasets}
