% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_folk.R
\name{as_folk}
\alias{as_folk}
\alias{as_folk.data.frame}
\alias{as_folk.data.table}
\alias{as_folk.list}
\alias{as_folk.tbl}
\title{Coerce data frames, lists, and more to data frames with as_folk subclass}
\usage{
as_folk(x, source = NULL)

\method{as_folk}{data.frame}(x, source = NULL)

\method{as_folk}{data.table}(x, source = NULL)

\method{as_folk}{list}(x, source = NULL)

\method{as_folk}{tbl}(x, source = NULL)
}
\arguments{
\item{x}{A data frame, list, matrix, or other object that could reasonably be
coerced to a data frame.}

\item{source}{A character vector denoting the source of the data,
e.g., "acs".}
}
\description{
\code{as_folk()} coerces an existing object, such as a data frame or data table,
to a data frame with a \code{folk_df} subclass.
}
\examples{
l <- list(a = c(1, 2), b = c(3, 4))
as_folk(l)
}
