% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_folk.R
\name{as_folk_task}
\alias{as_folk_task}
\alias{as_folk_task.folk_df}
\alias{as_folk_task.data.frame}
\alias{as_folk_task.data.table}
\title{Add a folk_task subclass to folk_df objects}
\usage{
as_folk_task(x, source = NULL, ...)

\method{as_folk_task}{folk_df}(x, source = NULL, ...)

\method{as_folk_task}{data.frame}(x, source = NULL, ...)

\method{as_folk_task}{data.table}(x, source = NULL, ...)
}
\arguments{
\item{x}{A folk_df object.}

\item{source}{A character vector denoting the source of the data,
e.g., "acs". If NULL, it will be inherited from the object passed to \code{x}.}

\item{...}{Other parameters passed onto methods.}
}
\description{
\code{as_folk_task()} adds the \code{folk_task} subclass to an existing folk_df object.
}
\examples{
l <- list(a = c(1, 2), b = c(3, 4))
df <- as_folk(l)
as_folk_task(df)
}
