% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folk-package.R
\docType{package}
\name{folk-package}
\alias{folk}
\alias{folk-package}
\title{folk: Datasets for Benchmarking Machine Learning Algorithms}
\description{
A suite of functions for downloading and deriving data from the US Census for benchmarking tasks. Inspired by the folktables Python package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/george-wood/epts}
  \item Report bugs at \url{https://github.com/george-wood/epts/issues}
}

}
\author{
\strong{Maintainer}: George Wood \email{georgebwood@gmail.com}

}
\keyword{internal}
