% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folk_df.R
\name{folk_df-class}
\alias{folk_df-class}
\alias{folk_df}
\title{\code{folk_df} class}
\description{
The \code{folk_df} class is a subclass of \code{\link[base:data.frame]{data.frame}},
created in order to allow method dispatch, object verification, and
consistency checking when using the folk package.
}
\section{Properties of \code{folk_df}}{


Objects of class \code{folk_df} have:
\itemize{
\item A \code{class} attribute of \code{c("folk_df", "data.frame")}.
\item A \code{source} attribute which can be queried to check the data source for
folk data. This attribute is used primarily to ensure that \code{folk_df}
objects are directed to the appropriate prediction tasks provided by the
folk package.
}
}

\section{Behavior of \code{folk_df}}{


The default behaviour of \code{folk_df} objects matches that of \link{data.frame}s.
The key difference is that:
\itemize{
\item If you run a function provided by the folk package, it will typically check
that an object has the class \code{folk_df} or is coercible to \code{folk_df}.
\item The \code{source} attribute will be queried to ensure that the \code{folk_df} object
can be used by the appropriate functions.
}
}

\seealso{
\code{as_folk()}
}
