% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{folk_example}
\alias{folk_example}
\title{Get path to folk example}
\usage{
folk_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
folk comes bundled with example files in its \code{inst/extdata}
directory. This function make them easy to access.
}
\examples{
folk_example()
folk_example("acs_example.csv")
}
