% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folk_task.R
\name{folk_task-class}
\alias{folk_task-class}
\alias{folk_task}
\title{\code{folk_task} class}
\description{
The \code{folk_task} class is a subclass of \code{folk_df} and
\code{\link[base:data.frame]{data.frame}}, created to represent a data object
that has been preprocessed using \code{set_task()}.
}
\section{Properties of \code{folk_task}}{


Objects of class \code{folk_task} have:
\itemize{
\item A \code{class} attribute of \code{c("folk_task", "folk_df", "data.frame")}.
\item A \code{source} attribute which can be queried to check the data source for
folk data.
}
}

\section{Behavior of \code{folk_task}}{


The default behaviour of \code{folk_task} objects matches that of \link{data.frame}s.
The key difference is that:
\itemize{
\item If you pass a \code{folk_task} object to \code{set_task()}, it will
generate an exception. This prevents the user from inadvertently setting
a task for a data object that has already been preprocessed for another,
possibly different task.
\item The \code{source} attribute will be queried to ensure that \code{folk_task} objects
are directed to the appropriate methods.
}

Inspired by the \code{tibble}\code{\link[tibble:tibble]{tibble::tibble()}} class, the \code{folk_task} class
implements printing and inspection functionality. This functionality is
designed to make it clear to the user which column is the target, which is
the group, and which are ordinary features for a particular task.
}

\seealso{
\code{as_folk()}, \code{print.folk_task()}
}
