% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_task}
\alias{set_task}
\title{Set a prediction task}
\usage{
set_task(data, task)
}
\arguments{
\item{data}{A data frame.}

\item{task}{A character string for the prediction task. See \code{\link[=show_tasks]{show_tasks()}}.}
}
\description{
Set a prediction task
}
