% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show_definition}
\alias{show_definition}
\title{Display definition of a folk task object}
\usage{
show_definition(data)
}
\arguments{
\item{data}{A folk_task object.}
}
\description{
Display definition of a folk task object
}
