% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{show_tasks}
\alias{show_tasks}
\title{Display available prediction tasks for a data object}
\usage{
show_tasks(data)
}
\arguments{
\item{data}{A data frame.}
}
\description{
Display available prediction tasks for a data object
}
